photogallerybackend/requirements.txt
asgiref==3.9.1
asn1crypto==1.5.1
Django==5.2.5
django-environ==0.12.0
django_pg8000==0.0.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.8.2
pg8000==1.31.4
pillow==11.3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
scramp==1.4.6
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
